# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Versatile Timesheet Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /timesheets:
    get:
      operationId: listTimesheets
      tags: [timesheets]
      summary: List timesheets with filters and pagination
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
        - name: employeeId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: weekStartDate
          in: query
          schema: { type: string }
        - name: managerId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of timesheets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimesheetList"
    post:
      operationId: createTimesheet
      tags: [timesheets]
      summary: Create or update a weekly timesheet draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimesheetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{timesheetId}:
    get:
      operationId: getTimesheet
      tags: [timesheets]
      summary: Get timesheet by ID
      parameters:
        - name: timesheetId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Timesheet with rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
    patch:
      operationId: updateTimesheet
      tags: [timesheets]
      summary: Update timesheet rows (draft only)
      parameters:
        - name: timesheetId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimesheetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{timesheetId}/submit:
    post:
      operationId: submitTimesheet
      tags: [timesheets]
      summary: Submit timesheet for approval
      parameters:
        - name: timesheetId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{timesheetId}/approve:
    post:
      operationId: approveTimesheet
      tags: [timesheets]
      summary: Approve a submitted timesheet
      parameters:
        - name: timesheetId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalAction"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{timesheetId}/reject:
    post:
      operationId: rejectTimesheet
      tags: [timesheets]
      summary: Reject a submitted timesheet
      parameters:
        - name: timesheetId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectionAction"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/bulk-action:
    post:
      operationId: bulkTimesheetAction
      tags: [timesheets]
      summary: Bulk approve or reject multiple timesheets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkTimesheetAction"
      responses:
        "200":
          description: Bulk action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkActionResult"
  /timesheets/copy-previous-week:
    post:
      operationId: copyPreviousWeek
      tags: [timesheets]
      summary: Copy previous week rows to new draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopyWeekRequest"
      responses:
        "200":
          description: New timesheet with copied rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"

  # NOTIFICATIONS
...
components:
  schemas:
...
    TimesheetRow:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        projectName:
          type: ["string", "null"]
        activityId:
          type: string
        activityName:
          type: ["string", "null"]
        monday:
          type: number
        tuesday:
          type: number
        wednesday:
          type: number
        thursday:
          type: number
        friday:
          type: number
        saturday:
          type: number
        sunday:
          type: number
        totalHours:
          type: number
        comments:
          type: ["string", "null"]
      required: [id, projectId, activityId, monday, tuesday, wednesday, thursday, friday, saturday, sunday, totalHours]
    TimesheetRowInput:
      type: object
      properties:
        projectId:
          type: string
        activityId:
          type: string
        monday:
          type: number
        tuesday:
          type: number
        wednesday:
          type: number
        thursday:
          type: number
        friday:
          type: number
        saturday:
          type: number
        sunday:
          type: number
        comments:
          type: ["string", "null"]
      required: [projectId, activityId, monday, tuesday, wednesday, thursday, friday, saturday, sunday]
    Timesheet:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        employeeName:
          type: ["string", "null"]
        weekStartDate:
          type: string
        weekEndDate:
          type: string
        status:
          type: string
          enum: [Draft, Submitted, Approved, Rejected]
        totalHours:
          type: number
        submittedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        approvedBy:
          type: ["string", "null"]
        approverName:
          type: ["string", "null"]
        rejectionComment:
          type: ["string", "null"]
        rows:
          type: array
          items:
            $ref: "#/components/schemas/TimesheetRow"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, employeeId, weekStartDate, weekEndDate, status, totalHours, rows, createdAt, updatedAt]
    TimesheetInput:
      type: object
      properties:
        employeeId:
          type: string
        weekStartDate:
          type: string
        rows:
          type: array
          items:
            $ref: "#/components/schemas/TimesheetRowInput"
      required: [employeeId, weekStartDate, rows]
    TimesheetUpdate:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/TimesheetRowInput"
      required: [rows]
    TimesheetList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Timesheet"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required: [data, total, page, pageSize]
    ApprovalAction:
      type: object
      properties:
        comment:
          type: ["string", "null"]
    RejectionAction:
      type: object
      properties:
        comment:
          type: string
      required: [comment]
    BulkTimesheetAction:
      type: object
      properties:
        timesheetIds:
          type: array
          items:
            type: string
        action:
          type: string
          enum: [approve, reject]
        comment:
          type: ["string", "null"]
      required: [timesheetIds, action]
    BulkActionResult:
      type: object
      properties:
        processed:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
      required: [processed, succeeded, failed]
    CopyWeekRequest:
      type: object
      properties:
        sourceWeekStartDate:
          type: string
        targetWeekStartDate:
          type: string
        employeeId:
          type: string
      required: [sourceWeekStartDate, targetWeekStartDate, employeeId]
...